EMKit has a few issues
It doesn't work with any transport other than
Email::Sender::Transport::Sendmail so if you don't
have a sendmail binary (e.g using Heroku) then it
will fail horribly the first time it tries to send
email. This branch is a everything I tried to make
it work and should not be used without picking out
the right bits.